feat: Flux2/ZImage/Qwen pipelines and GGUF improvements
Loaders:
- Add gguf_loader.py for combined GGUF file loading
- Update single_file.py, single_file_model.py, single_file_utils.py
- Add Ministral3 text encoder conversion support

Models:
- Update transformer_flux2.py with FP8Linear and enable_fp8_mode
- Update model_loading_utils.py with FP8 dtype preservation
- Update controlnet_z_image.py

Pipelines:
- Update pipeline_flux2.py
- Update pipeline_z_image.py and variants
- Update pipeline_qwenimage_edit_plus.py

Quantizers:
- Update gguf_quantizer.py and utils.py
- Add IQ2_XXS, IQ2_S, IQ3_S dequantization support
- Fix shape handling for combined GGUF files

Schedulers:
- Update flow_match_euler_discrete scheduler

Test files (untracked, for debugging):
- Various transformer/VAE/text encoder test scripts

+++ src/diffusers/models/controlnets/controlnet_z_image.py
@@ -692,7 +692,8 @@ class ZImageControlNetModel(ModelMixin, ConfigMixin, PeftAdapterMixin, FromOrigi
 
         # Match t_embedder output dtype to x for layerwise casting compatibility
         adaln_input = t.type_as(x)
-        x[torch.cat(x_inner_pad_mask)] = self.x_pad_token
+        # Cast x_pad_token to match x's dtype (handles FP8 transformer with float32 hidden states)
+        x[torch.cat(x_inner_pad_mask)] = self.x_pad_token.to(x.dtype)
         x = list(x.split(x_item_seqlens, dim=0))
         x_freqs_cis = list(self.rope_embedder(torch.cat(x_pos_ids, dim=0)).split([len(_) for _ in x_pos_ids], dim=0))
 
@@ -748,7 +749,8 @@ class ZImageControlNetModel(ModelMixin, ConfigMixin, PeftAdapterMixin, FromOrigi
 
         cap_feats = torch.cat(cap_feats, dim=0)
         cap_feats = self.cap_embedder(cap_feats)
-        cap_feats[torch.cat(cap_inner_pad_mask)] = self.cap_pad_token
+        # Cast cap_pad_token to match cap_feats dtype (handles FP8 transformer with float32 hidden states)
+        cap_feats[torch.cat(cap_inner_pad_mask)] = self.cap_pad_token.to(cap_feats.dtype)
         cap_feats = list(cap_feats.split(cap_item_seqlens, dim=0))
         cap_freqs_cis = list(
             self.rope_embedder(torch.cat(cap_pos_ids, dim=0)).split([len(_) for _ in cap_pos_ids], dim=0)
